use file based bytes iterator as default
